Move teleport handling into a separate class.
